remove: 등록한 트리 페이지에서만 사용되는 커스텀 훅 등록한 트리 페이지 하위 디렉토리로 이동
